fix(commands): add top-level folder command
